requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==3.1.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-debug-toolbar==3.1.1
importlib-metadata==1.7.0
pkg-resources==0.0.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.3.0
